+++ astropy/vo/client/vos_catalog.py
@@ -9,6 +9,7 @@ import fnmatch
 import json
 import os
 import re
+import socket
 import warnings
 
 from collections import defaultdict
@@ -841,6 +842,7 @@ def call_vo_service(service_type, catalog_db=None, pedantic=None,
         If VO service request fails.
 
     """
+    n_timed_out = 0
     catalogs = _get_catalogs(service_type, catalog_db, cache=cache,
                              verbose=verbose)
 
@@ -866,8 +868,13 @@ def call_vo_service(service_type, catalog_db=None, pedantic=None,
             return _vo_service_request(url, pedantic, kwargs, verbose=verbose)
         except Exception as e:
             vo_warn(W25, (url, str(e)))
+            if hasattr(e, 'reason') and isinstance(e.reason, socket.timeout):
+                n_timed_out += 1
 
-    raise VOSError('None of the available catalogs returned valid results.')
+    err_msg = 'None of the available catalogs returned valid results.'
+    if n_timed_out > 0:
+        err_msg += ' ({0} URL(s) timed out.)'.format(n_timed_out)
+    raise VOSError(err_msg)
 
 
 def list_catalogs(service_type, cache=True, verbose=True, **kwargs):